+++ src/Fibonacci/Fibonacci.java
@@ -1,3 +1,5 @@
+package Fibonacci;
+
 public class Fibonacci {
 
     public void fibonacci() {
